change priority_fee

+++ src/constants/mod.rs
@@ -55,6 +55,6 @@ pub mod accounts {
 pub mod trade {
     pub const JITO_TIP_AMOUNT: f64 = 0.0001;
     pub const DEFAULT_SLIPPAGE: u64 = 3000; // 30%
-    pub const DEFAULT_COMPUTE_UNIT_LIMIT: u32 = 78000;
-    pub const DEFAULT_COMPUTE_UNIT_PRICE: u64 = 500000;
+    pub const DEFAULT_COMPUTE_UNIT_LIMIT: f64 = 78000.0;
+    pub const DEFAULT_COMPUTE_UNIT_PRICE: f64 = 500000.0;
 }
